+++ src/directives/index.ts
@@ -4,6 +4,7 @@ import { applyBind, SLUG as BIND_SLUG } from './rz-bind';
 import { SLUG as FETCH_SLUG } from './rz-fetch';
 import { applyHtml, SLUG as HTML_SLUG } from './rz-html';
 import { SLUG as INSERT_SLUG } from './rz-insert';
+import { SLUG as ISLAND_SLUG } from './rz-island';
 import { applyModel, SLUG as MODEL_SLUG } from './rz-model';
 import { attachOn, SLUG as ON_SLUG } from './rz-on';
 import { SLUG as PUBLISH_SLUG } from './rz-publish';
@@ -11,7 +12,6 @@ import { SLUG as REFRESH_SLUG } from './rz-refresh';
 import { SLUG as STORE_SLUG } from './rz-store';
 import { applyText, SLUG as TEXT_SLUG } from './rz-text';
 import { SLUG as TUNE_SLUG } from './rz-tune';
-import { SLUG as USE_SLUG } from './rz-use';
 import { SLUG as WAKE_SLUG } from './rz-wake';
 
 // Directive types that are bound to DOM by attacher
@@ -27,11 +27,11 @@ export type ConfigDirectiveSlug =
   | typeof AUTOSAVE_SLUG
   | typeof FETCH_SLUG
   | typeof INSERT_SLUG
+  | typeof ISLAND_SLUG
   | typeof PUBLISH_SLUG
   | typeof REFRESH_SLUG
   | typeof STORE_SLUG
   | typeof TUNE_SLUG
-  | typeof USE_SLUG
   | typeof WAKE_SLUG;
 
 export type DirectiveSlug = DomDirectiveSlug | ConfigDirectiveSlug;
@@ -66,6 +66,7 @@ export { applyBind } from './rz-bind';
 export { cleanupFetch, handleFetch } from './rz-fetch';
 export { applyHtml } from './rz-html';
 export { getInsertConfig } from './rz-insert';
+export { getControllerName } from './rz-island';
 export { applyModel } from './rz-model';
 export { attachOn } from './rz-on';
 export { getPublishTopic } from './rz-publish';
@@ -73,5 +74,4 @@ export { attachRefresh } from './rz-refresh';
 export { getStoreName } from './rz-store';
 export { applyText } from './rz-text';
 export { getTuningStrategy } from './rz-tune';
-export { getControllerName } from './rz-use';
 export { processWake } from './rz-wake';