+++ src/directives/rz-island.ts
@@ -1,7 +1,7 @@
 import { splitInjection } from '../dom/utils';
 import { getDirective } from './prefix';
 
-export const SLUG = 'use' as const;
+export const SLUG = 'island' as const;
 
 export function getControllerName(el: HTMLElement): string | null {
   const raw = getDirective(el, SLUG);